Update files - Update add_to_readme() method to use jinja2 templating.     * Check if template file exists before accessing. - Update test. - Add README.md.template.

+++ auto_pb.py
@@ -10,6 +10,7 @@ import pathlib
 from pathlib import Path
 import re
 from shutil import rmtree
+from jinja2 import Template
 
 
 class ProjectBuilder:
@@ -183,11 +184,20 @@ class ProjectBuilder:
         elif self.proj_dir is None or not self.proj_dir.exists():
             raise FileNotFoundError('You need to create a README.md file.')
 
+        path_temp = Path.cwd() / 'README.md.template'
+
+        if not path_temp.exists():
+            raise FileNotFoundError('No README.md file template was found in '
+                                    'the current directory.')
+
+        readme_temp_str = path_temp.open('r').read()
+        readme_template = Template(readme_temp_str)
+
+        template_dict = {'project_name': self.proj_name,
+                         'author_name': self.author}
+        text_to_write = readme_template.render(template_dict)
+
         path = self.proj_dir / 'README.md'
-
-        text_to_write = f'# {self.proj_name}\nWelcome to {self.proj_name}!\n' \
-                        f'\n\n\nCreated by {self.author}.'
-
         with path.open('w') as readme:
             readme.write(text_to_write)
